requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
idna==3.4
pycodestyle==2.10.0
pydub==0.25.1
python-decouple==3.7
pytube==12.1.2
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
